# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: EffiPlat 全系统 API 文档
  version: "1.0"
  description: |
    EffiPlat 平台后端 API，涵盖用户、角色、权限、职责、环境、资产、服务、业务、Bug、认证等核心模块。
servers:
  - url: /api/v1
...
paths:
...
  /roles:
    get:
      tags: [roles]
      summary: 获取角色列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: name
          schema: { type: string }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleListResponse'
    post:
      tags: [roles]
      summary: 创建角色
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /roles/{roleId}:
    get:
      tags: [roles]
      summary: 获取角色详情
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: roleId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDetailsResponse'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [roles]
      summary: 更新角色
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: roleId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [roles]
      summary: 删除角色
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: roleId
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: 删除成功，无响应体
        '400':
          description: 角色有关联用户，无法删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Role:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    # ...（其余 schemas 省略，见前述 roles 示例）
...
    ErrorResponse:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        data: { type: [string, "null"] }
...
    RoleListResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string }
        data:
          type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Role' }
            total: { type: integer }
            page: { type: integer }
            pageSize: { type: integer }
    RoleResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string }
        data:
          $ref: '#/components/schemas/Role'
    RoleCreateRequest:
      type: object
      required: [name, description]
      properties:
        name: { type: string }
        description: { type: string }
    RoleUpdateRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
    RoleDetailsResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string }
        data:
          $ref: '#/components/schemas/Role'
...